More brightness stuff

+++ include/state.h
@@ -78,7 +78,7 @@ class State {
     std::string add = std::to_string(ADD);
     std::string top = std::to_string(TOP);
 
-    static unsigned int brightness;
+    static float brightness;
 
   private:
     unsigned int lastRender = 0;
